// backend/src/faturamento/type/DetailedBilling.type.ts
export interface TypeDetailedBilling {
  id: number | null
      cod_cli: string | null
      n_fatura: string | null
      cliente: string | null
      endereco: string | null
      bairro: string | null
      cidade: string | null
      uf: string | null
      cep: string | null
      cgc: string | null
      insc_esta: string | null
      insc_munic: string | null
      rps: string | null
      tp_nota: string | null
      vl_extenso: string | null
      valor_fatura: number | null
      valor_servicos: number | null
      dt_fatura: Date | null
      dt_vencimento: Date | null
      dt_entrada: Date | null
      iss_cobrar: string | null
      iss_valor: number | null
      iss_percentual: number | null
      iii_valor: number | null
      tributacao_msg: string | null
      observacao: string | null
      modalidade: string | null
      despachante: string | null
      valor_cif: number | null
      n_da: string | null
      n_di: string | null
      n_lote: string | null
      n_conhecimento: string | null
      n_documento: string | null
      tx_dolar: number | null
      nr_periodo_i: number | null
      nr_periodo_f: number | null
      dt_periodo_f: Date | null
      qt_volumes: number | null
      peso_bruto: number | null
      m3: number | null
      quantidade: number | null
      servico_id: string | null
      servico: string | null
      valor: number | null
      modalidade_txt: string | null
}